+++ mkdocs_monorepo_plugin/merger.py
@@ -51,18 +51,17 @@ class Merger:
             raise SystemExit(1)
 
         for alias, docs_dir in self.docs_dirs:
+            source_dir = docs_dir
             if len(alias) == 0:
-                source_dir = docs_dir
                 dest_dir = self.temp_docs_dir.name
             else:
-                source_dir = os.path.join(docs_dir, "docs")
                 split_alias = alias.split("/")
                 dest_dir = os.path.join(self.temp_docs_dir.name, *split_alias)
 
             if os.path.exists(source_dir):
                 copy_tree(source_dir, dest_dir)
                 for file_abs_path in Path(source_dir).rglob('*.md'):
-                    file_abs_path = str(file_abs_path) # python 3.5 compatibility
+                    file_abs_path = str(file_abs_path)  # python 3.5 compatibility
                     if os.path.isfile(file_abs_path):
                         file_rel_path = os.path.relpath(file_abs_path, source_dir)
                         dest = join(dest_dir, file_rel_path)